==> model/0_data_cleaning.ipynb <==
import json

import pandas as pd

def read_json(filename):
    with open(filename, encoding="UTF-8") as f:
        return list(json.load(f).items())
pd.DataFrame(read_json("input_data/train_set_ocr.json"))

def get_filename_hash(filename, separator):
    return filename[filename.rindex(separator) + 1:] if separator in filename else filename
get_filename_hash("./test_hashed/advertisement/00133f89-017d-44f5-a549-e11cf5b5221a.tiff", "//")

label_to_label_id =  pd.read_pickle('input_data/id2label_final.pkl')

filename_hash_to_label = {get_filename_hash(key, separator="/"): value for key, value in read_json("input_data/train_labels_final.json")}

## Add label and label_id

# ./test_hashed/.ipynb_checkpoints/18d1aa9b-c3cf-41e3-95db-bd668ada2000.tiff in input_data/train_set_ocr.json

train_dataset = pd.DataFrame(read_json("input_data/train_set_ocr.json"), columns="filename text".split()).rename({"index": "filename"}, axis=1)
train_dataset["filename"] = train_dataset["filename"].apply(get_filename_hash, args=["/"])

not_found = []
for filename in train_dataset["filename"]:
    if filename not in filename_hash_to_label:
        not_found.append(filename)
print(f"Not found: {len(not_found)}")
train_dataset = train_dataset[~train_dataset["filename"].isin(not_found)]

train_dataset["label"] = train_dataset["filename"].apply(lambda f: filename_hash_to_label[f]).astype("category")
train_dataset["label_id"] = train_dataset["label"].apply(lambda l: label_to_label_id[l]).astype("category")
train_dataset.info()

assert all(label in label_to_label_id for label in train_dataset["label"].to_list())

train_dataset.head()

train_dataset[["label", "label_id"]].value_counts()

train_dataset.info()

## Labels number

print(len(train_dataset["label"].unique()))

## Language Split

import string

def tokenize(text: str):
    text = text.lower()

    punctuation = set(string.punctuation) - {"-"}
    for punctuation_mark in punctuation:
        text = text.replace(punctuation_mark, "")
    text = "".join(c for c in text if not c.isdigit())

    word_list = [word.strip().strip("-") for word in text.split()]
    word_list = [word for word in word_list if word]
    return " ".join(word_list)

train_dataset["text"] = train_dataset["text"].apply(tokenize)
train_dataset.tail()

print("Mean number of characters:", train_dataset["text"].apply(lambda s: len(s)).mean())
print("Mean number of words:", train_dataset["text"].apply(lambda s: len(s.split())).mean())

train_dataset.to_csv("train_input_data_pl_and_eng.csv", index=False)
first_english_index = train_dataset["filename"].to_list().index("a64b06e6-8f0a-4fd7-ba12-0779c5560d9a.tiff")
train_dataset_pl = train_dataset.iloc[:first_english_index]
train_dataset_eng = train_dataset.iloc[first_english_index:]

train_dataset_pl.to_csv("train_input_data_pl.csv", index=False)
train_dataset_eng.to_csv("train_input_data_eng.csv", index=False)

print("Polish language in training dataset", train_dataset_pl.shape[0] / train_dataset.shape[0])
