==> model/2_detect_language.ipynb <==
import pandas as pd
from py3langid.langid import LanguageIdentifier, MODEL_FILE
from langdetect import detect, detect_langs

df = pd.read_csv("train_input_data_pl_and_eng.csv")
df = df.dropna()
df['words_count'] = df["text"].apply(lambda text: len(text.split()))

identifier = LanguageIdentifier.from_pickled_model(MODEL_FILE)
identifier.set_languages(['pl', 'en'])
df["lang"] = df["text"].apply(lambda text: identifier.classify(text)[0])
df
